Add some documentation to explain how score_list works.
Add a sanity check on score_list and fix two bugs discovered because of it.
One was in how skip_first_couple was used, and the other was with
limit_edit_distance.  Limit_edit_distance will now never return a result
which involved more edits than the limit passed in as that result could
be non-optimal.